Задание 4

+++ homework/task4.py
@@ -1,4 +1,4 @@
 print('Для нахождения площади квадрата введите длину стороны')
 square_side = input()
-square_side = int(square_side)
+square_side = float(square_side)
 print('Площадь квадрата равна', square_side**2)
